fix: structure refactoring
