+++ src/Product/View/Info.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { FormattedNumber } from 'react-intl';
-import Code from '../common/ProductCode';
-import { HiddenLg } from '../common/Hidden';
+import Code from './Code';
+import { HiddenLg } from '../../common/Responsive';
 
 const Info = styled.div`
   display: flex;
