+++ src/Product/View/Slider.jsx
