+++ src/Product/View/index.jsx
@@ -1,14 +1,13 @@
 import React from 'react';
 import Card from './Card';
-import Accordion from '../Accordion';
+import Accordion from './Accordion';
 import Recomendations from './Recomendations';
 import Links from './Links';
 
-export default () => (
-  <main>
+export default () =>
+  (<main>
     <Card />
     <Accordion />
     <Recomendations />
     <Links />
-  </main>
-);
+  </main>);
