+++ src/ProductList.jsx
@@ -4,10 +4,7 @@ import en from 'react-intl/locale-data/en';
 import ru from 'react-intl/locale-data/ru';
 import Helmet from 'react-helmet';
 import Header from './Header';
-import Filters from './Filters';
-import ProductCategory from './Product/ProductCategory';
-import Divider from './common/Divider';
-import More from './Product/ShowMore';
+import ProductList from './Product/List';
 import Footer from './Footer';
 
 addLocaleData([...en, ...ru]);
@@ -20,13 +17,7 @@ export default () =>
         <meta name="description" content="Men's clothing | Burberry" />
       </Helmet>
       <Header />
-      <Filters />
-      <ProductCategory heading="Heritage Trench Coats" />
-      <div className="container">
-        <Divider />
-      </div>
-      <ProductCategory heading="Single Breasted Trench Coats" />
-      <More viewed="8" summ="17" count="9" />
+      <ProductList />
       <Footer />
     </div>
   </IntlProvider>);
