+++ src/index.js
@@ -5,8 +5,8 @@ import ReactDOM from 'react-dom';
 import 'flexboxgrid2/dist/flexboxgrid.css';
 import 'normalize.css';
 import registerServiceWorker from './registerServiceWorker';
-import Products from './Products';
+import Product from './Product';
 import './index.css';
 
-ReactDOM.render(<Products />, document.getElementById('root'));
+ReactDOM.render(<Product />, document.getElementById('root'));
 registerServiceWorker();
